digital-signage-cms/ai-services/requirements.txt
openai 
python-dotenv 
